organised image and svg files
