+++ src/assets/images/landing/cat-circle.png
