+++ src/assets/images/landing/dog-circle.png
