+++ src/assets/images/landing/horse-circle.png
