+++ src/assets/images/search-all/hero.jpg
